add order service

+++ cart-service/src/main/java/com/perepalacin/cart_service/service/CartService.java
@@ -3,7 +3,9 @@ package com.perepalacin.cart_service.service;
 import com.perepalacin.cart_service.client.ProductServiceClient;
 import com.perepalacin.cart_service.entity.dao.Cart;
 import com.perepalacin.cart_service.entity.dao.CartItem;
-import com.perepalacin.cart_service.entity.dao.ProductDao;
+import com.perepalacin.cart_service.entity.dto.CartDto;
+import com.perepalacin.cart_service.entity.dto.CartItemDto;
+import com.perepalacin.cart_service.entity.dto.ProductDto;
 import com.perepalacin.cart_service.entity.dto.UserDetailsDto;
 import com.perepalacin.cart_service.exceptions.ProductNotFoundException;
 import com.perepalacin.cart_service.exceptions.UnauthorizedException;
@@ -11,34 +13,56 @@ import com.perepalacin.cart_service.repository.CartItemRepository;
 import com.perepalacin.cart_service.repository.CartRepository;
 import com.perepalacin.cart_service.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class CartService {
 
     private final CartRepository cartRepository;
     private final ProductServiceClient productServiceClient;
     private final CartItemRepository cartItemRepository;
 
-    public Cart getUserCart () {
+    public CartDto getUserCart () {
         UserDetailsDto userDetails = JwtUtil.getCredentialsFromToken();
         if (userDetails.getUserId() == null) {
             throw new UnauthorizedException();
         }
+        Cart cart = cartRepository.getByUserId(userDetails.getUserId()).orElse(null);
 
-        return cartRepository.getByUserId(userDetails.getUserId()).orElse(null);
+        if (cart != null) {
+            List<CartItemDto> cartItems = this.getCartItemDetails(cart);
+            return CartDto.builder().id(cart.getId()).items(cartItems).build();
+        } else {
+            return CartDto.builder().build();
+        }
     }
 
+    public CartDto getCartByUserId (final UUID userId) {
+        UserDetailsDto userDetails = JwtUtil.getCredentialsFromToken();
+        if (userDetails.getUserId() == null || !userDetails.getUserId().equals(userId)) {
+            throw new UnauthorizedException();
+        }
+        Cart cart = cartRepository.getByUserId(userId).orElse(null);
+
+        if (cart != null) {
+            List<CartItemDto> cartItems = this.getCartItemDetails(cart);
+            return CartDto.builder().id(cart.getId()).items(cartItems).build();
+        } else {
+            return CartDto.builder().build();
+        }
+    }
+
+
     @Transactional
-    public Cart addItemToCart (final Long productId, final Integer quantity) {
+    public void addItemToCart (final Long productId, final Integer quantity) {
         UserDetailsDto userDetails = JwtUtil.getCredentialsFromToken();
         if (userDetails.getUserId() == null) {
             throw new UnauthorizedException();
@@ -60,15 +84,15 @@ public class CartService {
             if (Objects.equals(cart.getItems().get(i).getProductId(), productId)) {
                 cart.getItems().get(i).setQuantity(cart.getItems().get(i).getQuantity() + quantity);
                 cartRepository.save(cart);
-                return cart;
+                return;
             }
         }
 
-        ResponseEntity<ProductDao> productResponse = productServiceClient.getProductById(productId);
-        if (HttpStatus.OK.equals(productResponse.getStatusCode())) {
+        ResponseEntity<ProductDto> productResponse = productServiceClient.getProductById(productId);
+        if (productResponse.getStatusCode().is2xxSuccessful()) {
             cart.getItems().add(CartItem.builder().productId(productId).quantity(quantity).cart(cart).build());
             cartRepository.save(cart);
-            return cart;
+            return;
         } else if (HttpStatus.NOT_FOUND.equals(productResponse.getStatusCode())) {
             throw new ProductNotFoundException("Product with id: " + productId + " not found");
         } else {
@@ -76,7 +100,7 @@ public class CartService {
         }
     }
 
-    public CartItem editCartItemQuantity (final Long cartItemId, final Integer quantity) {
+    public CartItemDto editCartItemQuantity (final Long cartItemId, final Integer quantity) {
         UserDetailsDto userDetails = JwtUtil.getCredentialsFromToken();
         if (userDetails.getUserId() == null) {
             throw new UnauthorizedException();
@@ -86,13 +110,28 @@ public class CartService {
                 () ->  new ProductNotFoundException("Cart item with id: " + cartItemId + " not found")
         );
 
+        ResponseEntity<ProductDto> productResponse = productServiceClient.getProductById(cartItem.getProductId());
+
+        if (!productResponse.getStatusCode().is2xxSuccessful() || productResponse.getBody() == null) {
+            throw new RuntimeException("Something went wrong while getting the product details, please try again later");
+        }
+
         if (quantity == 0) {
             cartItemRepository.delete(cartItem);
             return null;
         } else {
             cartItem.setQuantity(quantity);
             cartItemRepository.save(cartItem);
-            return cartItem;
+            return CartItemDto.builder()
+                    .id(cartItem.getId())
+                    .quantity(quantity)
+                    .price(productResponse.getBody().getPrice())
+                    .stock(productResponse.getBody().getStock())
+                    .name(productResponse.getBody().getName())
+                    .imageUrl(productResponse.getBody().getImageUrl())
+                    .publicUrl(productResponse.getBody().getPublicUrl())
+                    .productId(cartItem.getProductId())
+                    .build();
         }
     }
 
@@ -116,4 +155,45 @@ public class CartService {
 
         cartRepository.delete(cart);
     }
+
+    private List<CartItemDto> getCartItemDetails (Cart cart) {
+        List<Long> productIds = new ArrayList<>();
+        cart.getItems().forEach(item -> productIds.add(item.getProductId()));
+
+        ResponseEntity<List<ProductDto>> productResponse = productServiceClient.getListOfProductsById(productIds);
+
+        log.debug(productResponse.getStatusCode().toString());
+        log.info(String.valueOf(productResponse.getBody().size()));
+        if (productResponse.getStatusCode().is2xxSuccessful() && productResponse.getBody() != null) {
+            List<CartItemDto> cartItems = new ArrayList<>();
+            productResponse.getBody().forEach(product -> {
+
+                Optional<CartItem> foundItem = cart.getItems().stream()
+                        .filter(item -> item.getProductId() != null && item.getProductId().equals(product.getId()))
+                        .findFirst();
+
+                if (foundItem.isEmpty()) {
+                    throw new RuntimeException("There seems to be an error with the items of your cart, please try again later.");
+                }
+
+                cartItems.add(CartItemDto.builder()
+                        .id(foundItem.get().getId())
+                        .quantity(foundItem.get().getQuantity())
+                        .productId(product.getId())
+                        .name(product.getName())
+                        .price(product.getPrice())
+                        .stock(product.getStock())
+                        .publicUrl(product.getPublicUrl())
+                        .productId(product.getId())
+                        .imageUrl(product.getImageUrl())
+                        .build());
+                return;
+            });
+
+            return cartItems;
+        } else {
+            throw new RuntimeException("Something went wrong while fetching the products in your cart, please try again later.");
+        }
+    }
 }
+
